chore: separate connection manager from deployment
