+++ client/src/app/App.js
@@ -2422,6 +2422,7 @@ export class App extends PureComponent {
                         settings={ this.getGlobal('settings') }
                         backend={ this.getGlobal('backend') }
                         config={ this.getGlobal('config') }
+                        connectionManager={ this.getGlobal('connectionManager') }
                         deployment={ this.getGlobal('deployment') }
                         startInstance={ this.getGlobal('startInstance') }
                         zeebeApi={ this.getGlobal('zeebeAPI') }
