+++ client/src/app/__tests__/IntegrationSpec.js
@@ -20,6 +20,7 @@ import {
   Backend,
   Cache,
   Config,
+  ConnectionManager,
   Deployment,
   Dialog,
   FileSystem,
@@ -147,6 +148,7 @@ function createApp(options = {}) {
   let globals = {
     backend: new Backend(),
     config: new Config(),
+    connectionManager: new ConnectionManager(),
     deployment: new Deployment(),
     dialog: new Dialog(),
     fileSystem: new FileSystem(),
