+++ client/src/app/tabs/MultiSheetTab.js
@@ -412,6 +412,7 @@ export class MultiSheetTab extends CachedComponent {
             settings={ this.props.settings }
             backend={ this.props.backend }
             config={ this.props.config }
+            connectionManager={ this.props.connectionManager }
             deployment={ this.props.deployment }
             startInstance={ this.props.startInstance }
             zeebeApi={ this.props.zeebeApi }
