+++ client/src/app/zeebe/__tests__/DeploymentSpec.js
@@ -399,178 +399,6 @@ describe('Deployment', function() {
     expect(gatewayVersion).to.equal('foo');
   });
 
-
-  describe('#getEndpoints', function() {
-
-    it('should return empty array when settings returns null', function() {
-
-      // given
-      const settings = {
-        get: sinon.stub().returns(null)
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.deep.equal([]);
-    });
-
-
-    it('should return empty array when settings returns undefined', function() {
-
-      // given
-      const settings = {
-        get: sinon.stub().returns(undefined)
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.deep.equal([]);
-    });
-
-
-    it('should return empty array when settings returns non-array value', function() {
-
-      // given
-      const settings = {
-        get: sinon.stub().returns('invalid string')
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.deep.equal([]);
-    });
-
-
-    it('should return empty array when settings returns object', function() {
-
-      // given
-      const settings = {
-        get: sinon.stub().returns({ corrupted: 'object' })
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.deep.equal([]);
-    });
-
-
-    it('should filter out connections without id', function() {
-
-      // given
-      const connections = [
-        {
-          id: 'connection-1',
-          name: 'Valid Connection 1',
-          url: 'http://localhost:8080'
-        },
-        {
-          name: 'Connection without ID',
-          url: 'http://localhost:8081'
-        },
-        {
-          id: '',
-          name: 'Connection with empty ID',
-          url: 'http://localhost:8082'
-        },
-        {
-          id: 'connection-2',
-          name: 'Valid Connection 2',
-          url: 'http://localhost:8083'
-        }
-      ];
-
-      const settings = {
-        get: sinon.stub().returns(connections)
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.have.length(2);
-      expect(endpoints[0]).to.have.property('id', 'connection-1');
-      expect(endpoints[1]).to.have.property('id', 'connection-2');
-    });
-
-
-    it('should return valid connections array from settings', function() {
-
-      // given
-      const connections = [
-        {
-          id: 'connection-1',
-          name: 'Test Connection 1',
-          url: 'http://localhost:8080'
-        },
-        {
-          id: 'connection-2',
-          name: 'Test Connection 2',
-          url: 'http://localhost:8081'
-        }
-      ];
-
-      const settings = {
-        get: sinon.stub().returns(connections)
-      };
-
-      const deployment = createDeployment({ settings });
-
-      // when
-      const endpoints = deployment.getEndpoints();
-
-      // then
-      expect(endpoints).to.deep.equal(connections);
-    });
-  });
-
-
-  describe('#getConnectionForTab', function() {
-
-    it('should return NO_CONNECTION without tab', async function() {
-
-      // given
-      const deployment = createDeployment();
-
-      // when
-      const result = await deployment.getConnectionForTab();
-
-      // then
-      expect(result.id).to.equal('NO_CONNECTION');
-    });
-
-
-    it('should return NO_CONNECTION without tab.file', async function() {
-
-      // given
-      const deployment = createDeployment();
-
-      // when
-      const result = await deployment.getConnectionForTab({ file: undefined });
-
-      // then
-      expect(result.id).to.equal('NO_CONNECTION');
-    });
-
-  });
-
 });
 
 class Mock {
