+++ client/src/plugins/process-applications/__tests__/ProcessApplicationsStartInstancePluginSpec.js
@@ -19,7 +19,6 @@ import { mount } from 'enzyme';
 import ProcessApplicationsStartInstancePlugin from '../ProcessApplicationsStartInstancePlugin';
 
 import { Slot, SlotFillRoot } from '../../../app/slot-fill';
-import { DEFAULT_ENDPOINT } from '../../../app/zeebe/Deployment';
 
 import { Deployment, StartInstance, ZeebeAPI } from '../../../app/__tests__/mocks';
 
@@ -154,17 +153,9 @@ function createProcessApplicationsStartInstancePlugin(props = {}) {
   const {
     _getGlobal = (name) => {
       if (name === 'deployment') {
-        return new Deployment({
-          async getConnectionForTab(file) {
-            return DEFAULT_ENDPOINT;
-          }
-        });
+        return new Deployment();
       } else if (name === 'startInstance') {
-        return new StartInstance({
-          async getConnectionForTab(file) {
-            return {};
-          }
-        });
+        return new StartInstance();
       } else if (name === 'zeebeAPI') {
         return new ZeebeAPI();
       }
