+++ client/src/plugins/zeebe-plugin/deployment-plugin/__tests__/DeploymentPluginOverlaySpec.js
@@ -16,8 +16,6 @@ import { waitFor } from '@testing-library/react';
 
 import { mount } from 'enzyme';
 
-import EventEmitter from 'events';
-
 import DeploymentPluginOverlay from '../DeploymentPluginOverlay';
 
 import { TARGET_TYPES } from '../../../../remote/ZeebeAPI';
@@ -40,9 +38,12 @@ describe('DeploymentPluginOverlay', function() {
       // given
       const endpoint = createMockEndpoint();
 
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
-        getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -50,6 +51,7 @@ describe('DeploymentPluginOverlay', function() {
       const { Form, getProps: getFormProps } = createMockDeploymentConfigForm();
 
       createDeploymentPluginOverlay({
+        connectionManager,
         deployment,
         DeploymentConfigForm: Form,
         displayNotification: displayNotificationSpy
@@ -87,6 +89,10 @@ describe('DeploymentPluginOverlay', function() {
       // given
       const endpoint = createMockEndpoint();
 
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult({
           success: false,
@@ -94,7 +100,6 @@ describe('DeploymentPluginOverlay', function() {
             details: 'foo'
           }
         }))),
-        getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -104,6 +109,7 @@ describe('DeploymentPluginOverlay', function() {
       const logSpy = sinon.spy();
 
       createDeploymentPluginOverlay({
+        connectionManager,
         deployment,
         DeploymentConfigForm: Form,
         displayNotification: displayNotificationSpy,
@@ -155,15 +161,21 @@ describe('DeploymentPluginOverlay', function() {
 
       const mockDeploymentResult = createMockDeploymentResult();
 
-      const deployment = new MockDeployment({
-        deploy: sinon.spy(() => Promise.resolve(mockDeploymentResult)),
-        on: sinon.spy(),
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
       });
+
+      const deployment = new MockDeployment({
+        deploy: sinon.spy(() => Promise.resolve(mockDeploymentResult)),
+        on: sinon.spy()
+      });
+
+
 
       const triggerActionSpy = sinon.spy();
 
       createDeploymentPluginOverlay({
+        connectionManager,
         deployment,
         triggerAction: triggerActionSpy
       });
@@ -209,15 +221,19 @@ describe('DeploymentPluginOverlay', function() {
         response: createMockDeploymentErrorResponse()
       });
 
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(mockDeploymentResult)),
-        on: sinon.spy(),
-        getConnectionForTab: () => Promise.resolve(endpoint)
+        on: sinon.spy()
       });
 
       const triggerActionSpy = sinon.spy();
 
       createDeploymentPluginOverlay({
+        connectionManager,
         deployment,
         triggerAction: triggerActionSpy
       });
@@ -265,12 +281,12 @@ describe('DeploymentPluginOverlay', function() {
       // when
       const endpoint = createMockEndpoint();
 
-      const deployment = new MockDeployment({
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
       createDeploymentPluginOverlay({
-        deployment,
+        connectionManager,
         renderHeader: <div id="custom-header" />
       });
 
@@ -288,12 +304,12 @@ describe('DeploymentPluginOverlay', function() {
       // when
       const endpoint = createMockEndpoint();
 
-      const deployment = new MockDeployment({
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
       createDeploymentPluginOverlay({
-        deployment,
+        connectionManager,
         renderSubmit: <div id="custom-submit" />
       });
 
@@ -311,9 +327,12 @@ describe('DeploymentPluginOverlay', function() {
       // given
       const endpoint = createMockEndpoint();
 
-      const deployment = new MockDeployment({
-        deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
+      const deployment = new MockDeployment({
+        deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult()))
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -328,6 +347,7 @@ describe('DeploymentPluginOverlay', function() {
       ];
 
       createDeploymentPluginOverlay({
+        connectionManager,
         deployment,
         DeploymentConfigForm: Form,
         displayNotification: displayNotificationSpy,
@@ -366,9 +386,12 @@ describe('DeploymentPluginOverlay', function() {
       // given
       const endpoint = createMockEndpoint();
 
-      const deployment = new MockDeployment({
-        deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
+      const deployment = new MockDeployment({
+        deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult()))
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -377,6 +400,7 @@ describe('DeploymentPluginOverlay', function() {
 
       createDeploymentPluginOverlay({
         deployment,
+        connectionManager,
         DeploymentConfigForm: Form,
         displayNotification: displayNotificationSpy,
         getSuccessNotification: () => ({
@@ -417,14 +441,17 @@ describe('DeploymentPluginOverlay', function() {
       // given
       const endpoint = createMockEndpoint();
 
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult({
           success: false,
           response: {
             details: 'foo'
           }
-        }))),
-        getConnectionForTab: () => Promise.resolve(endpoint)
+        })))
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -433,6 +460,7 @@ describe('DeploymentPluginOverlay', function() {
 
       createDeploymentPluginOverlay({
         deployment,
+        connectionManager,
         DeploymentConfigForm: Form,
         displayNotification: displayNotificationSpy,
         getErrorNotification: () => ({
@@ -477,38 +505,16 @@ class Mock {
   }
 }
 
-class MockConnectionChecker extends Mock {
-  constructor(overrides = {}) {
-    super(overrides);
-
-    this.eventEmitter = new EventEmitter();
-  }
-
-  on = sinon.spy((...args) => {
-    return this.eventEmitter.on(...args);
-  });
-
-  off = sinon.spy((...args) => {
-    return this.eventEmitter.off(...args);
-  });
-
-  emit = sinon.spy((...args) => {
-    return this.eventEmitter.emit(...args);
-  });
-
-  updateConfig = sinon.spy();
-
-  startChecking = sinon.spy();
-
-  stopChecking = sinon.spy();
-}
-
 class MockDeployment extends Mock {
   deploy() {}
 
   off() {}
 
   on() {}
+}
+
+class MockConnectionManager extends Mock {
+  getConnectionForTab() {}
 }
 
 class MockConfigValidator extends Mock {
@@ -611,7 +617,7 @@ function createDeploymentPluginOverlay(props = {}) {
   const {
     activeTab = DEFAULT_ACTIVE_TAB,
     anchor = new MockAnchor(),
-    connectionChecker = new MockConnectionChecker(),
+    connectionManager = new MockConnectionManager(),
     deployment = new MockDeployment(),
     deploymentConfigValidator = MockConfigValidator,
     displayNotification = () => {},
@@ -630,7 +636,7 @@ function createDeploymentPluginOverlay(props = {}) {
     <DeploymentPluginOverlay
       activeTab={ activeTab }
       anchor={ anchor }
-      connectionChecker={ connectionChecker }
+      connectionManager={ connectionManager }
       deployment={ deployment }
       deploymentConfigValidator={ deploymentConfigValidator }
       displayNotification={ displayNotification }
