+++ client/src/plugins/zeebe-plugin/start-instance-plugin/__tests__/StartInstancePluginOverlaySpec.js
@@ -16,8 +16,6 @@ import { waitFor } from '@testing-library/react';
 
 import { mount } from 'enzyme';
 
-import EventEmitter from 'events';
-
 import StartInstancePluginOverlay from '../StartInstancePluginOverlay';
 
 import { TARGET_TYPES } from '../../../../remote/ZeebeAPI';
@@ -46,10 +44,10 @@ describe('StartInstancePluginOverlay', function() {
   it('should render start instance config (deployment config valid, no connection check result)', async function() {
 
     // when
-    const deploymentConfig = createMockEndpoint();
+    const connection = createMockConnection();
 
-    const deployment = new MockDeployment({
-      getConnectionForTab: () => Promise.resolve(deploymentConfig)
+    const connectionManager = new MockConnectionManager({
+      getConnectionForTab: () => Promise.resolve(connection)
     });
 
     const startInstanceConfig = createMockStartInstanceConfig();
@@ -59,7 +57,7 @@ describe('StartInstancePluginOverlay', function() {
     });
 
     createStartInstancePluginOverlay({
-      deployment,
+      connectionManager,
       startInstance,
       StartInstanceConfigForm: createMockStartInstanceConfigForm().Form
     });
@@ -78,10 +76,14 @@ describe('StartInstancePluginOverlay', function() {
       it('should submit form (success)', async function() {
 
         // given
-        const endpoint = createMockEndpoint();
+        const endpoint = createMockConnection();
 
         const deployment = new MockDeployment({
           deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
+          getConnectionForTab: () => Promise.resolve(endpoint)
+        });
+
+        const connectionManager = new MockConnectionManager({
           getConnectionForTab: () => Promise.resolve(endpoint)
         });
 
@@ -98,6 +100,7 @@ describe('StartInstancePluginOverlay', function() {
 
         createStartInstancePluginOverlay({
           deployment,
+          connectionManager,
           displayNotification: displayNotificationSpy,
           startInstance,
           StartInstanceConfigForm: Form
@@ -132,10 +135,13 @@ describe('StartInstancePluginOverlay', function() {
       it('should submit form (no success)', async function() {
 
         // given
-        const endpoint = createMockEndpoint();
+        const endpoint = createMockConnection();
 
         const deployment = new MockDeployment({
-          deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
+          deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult()))
+        });
+
+        const connectionManager = new MockConnectionManager({
           getConnectionForTab: () => Promise.resolve(endpoint)
         });
 
@@ -157,6 +163,7 @@ describe('StartInstancePluginOverlay', function() {
 
         createStartInstancePluginOverlay({
           deployment,
+          connectionManager,
           displayNotification: displayNotificationSpy,
           startInstance,
           StartInstanceConfigForm: Form
@@ -198,7 +205,7 @@ describe('StartInstancePluginOverlay', function() {
     it('should emit event (success)', async function() {
 
       // given
-      const endpoint = createMockEndpoint();
+      const endpoint = createMockConnection();
 
       const mockDeploymentResult = createMockDeploymentResult();
 
@@ -257,7 +264,7 @@ describe('StartInstancePluginOverlay', function() {
     it('should emit event (error)', async function() {
 
       // given
-      const endpoint = createMockEndpoint();
+      const endpoint = createMockConnection();
 
       const mockDeploymentResult = createMockDeploymentResult({
         success: false,
@@ -326,10 +333,10 @@ describe('StartInstancePluginOverlay', function() {
     it('should render custom start instance header', async function() {
 
       // when
-      const deploymentConfig = createMockEndpoint();
+      const connection = createMockConnection();
 
-      const deployment = new MockDeployment({
-        getConnectionForTab: () => Promise.resolve(deploymentConfig)
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(connection)
       });
 
       const startInstanceConfig = createMockStartInstanceConfig();
@@ -339,7 +346,7 @@ describe('StartInstancePluginOverlay', function() {
       });
 
       createStartInstancePluginOverlay({
-        deployment,
+        connectionManager,
         renderStartInstanceHeader: <div id="custom-start-instance-header" />,
         startInstance
       });
@@ -356,10 +363,10 @@ describe('StartInstancePluginOverlay', function() {
     it('should render custom start instance submit', async function() {
 
       // when
-      const deploymentConfig = createMockEndpoint();
+      const connection = createMockConnection();
 
-      const deployment = new MockDeployment({
-        getConnectionForTab: () => Promise.resolve(deploymentConfig)
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(connection)
       });
 
       const startInstanceConfig = createMockStartInstanceConfig();
@@ -369,7 +376,7 @@ describe('StartInstancePluginOverlay', function() {
       });
 
       createStartInstancePluginOverlay({
-        deployment,
+        connectionManager,
         renderStartInstanceSubmit: <div id="custom-start-instance-submit" />,
         startInstance
       });
@@ -386,7 +393,7 @@ describe('StartInstancePluginOverlay', function() {
     it('should deploy custom resources', async function() {
 
       // given
-      const endpoint = createMockEndpoint();
+      const endpoint = createMockConnection();
 
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult({
@@ -407,6 +414,9 @@ describe('StartInstancePluginOverlay', function() {
             ]
           }
         }))),
+      });
+
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
@@ -434,6 +444,7 @@ describe('StartInstancePluginOverlay', function() {
 
       createStartInstancePluginOverlay({
         deployment,
+        connectionManager,
         displayNotification: displayNotificationSpy,
         getResourceConfigs: () => resourceConfigs,
         startInstance,
@@ -481,10 +492,13 @@ describe('StartInstancePluginOverlay', function() {
     it('should display custom success notification', async function() {
 
       // given
-      const endpoint = createMockEndpoint();
+      const endpoint = createMockConnection();
 
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
+      });
+
+      const connectionManager = new MockConnectionManager({
         getConnectionForTab: () => Promise.resolve(endpoint)
       });
 
@@ -501,6 +515,7 @@ describe('StartInstancePluginOverlay', function() {
 
       createStartInstancePluginOverlay({
         deployment,
+        connectionManager,
         displayNotification: displayNotificationSpy,
         getSuccessNotification: () => ({
           title: 'Custom success notification',
@@ -540,11 +555,14 @@ describe('StartInstancePluginOverlay', function() {
     it('should display custom error notification', async function() {
 
       // given
-      const endpoint = createMockEndpoint();
+      const connection = createMockConnection();
 
       const deployment = new MockDeployment({
         deploy: sinon.spy(() => Promise.resolve(createMockDeploymentResult())),
-        getConnectionForTab: () => Promise.resolve(endpoint)
+      });
+
+      const connectionManager = new MockConnectionManager({
+        getConnectionForTab: () => Promise.resolve(connection)
       });
 
       const displayNotificationSpy = sinon.spy();
@@ -564,6 +582,7 @@ describe('StartInstancePluginOverlay', function() {
       const { Form, getProps: getFormProps } = createMockStartInstanceConfigForm();
 
       createStartInstancePluginOverlay({
+        connectionManager,
         deployment,
         displayNotification: displayNotificationSpy,
         getErrorNotification: () => ({
@@ -587,7 +606,7 @@ describe('StartInstancePluginOverlay', function() {
       });
 
       expect(startInstance.startInstance).to.have.been.calledWith({
-        endpoint,
+        endpoint: connection,
         context: 'startInstancePlugin',
         ...startInstanceConfig
       });
@@ -609,42 +628,16 @@ class Mock {
   }
 }
 
-class MockConnectionChecker extends Mock {
-  constructor(overrides = {}) {
-    super(overrides);
-
-    this.eventEmitter = new EventEmitter();
-  }
-
-  on = sinon.spy((...args) => {
-    return this.eventEmitter.on(...args);
-  });
-
-  off = sinon.spy((...args) => {
-    return this.eventEmitter.off(...args);
-  });
-
-  emit = sinon.spy((...args) => {
-    return this.eventEmitter.emit(...args);
-  });
-
-  updateConfig = sinon.spy();
-
-  startChecking = sinon.spy();
-
-  stopChecking = sinon.spy();
-}
-
 class MockDeployment extends Mock {
   deploy() {}
-
-  getConnectionForTab() {}
 
   off() {}
 
   on() {}
+}
 
-  setConfigForFile() {}
+class MockConnectionManager extends Mock {
+  getConnectionForTab() {}
 }
 
 class MockConfigValidator extends Mock {
@@ -743,7 +736,7 @@ const DEFAULT_ACTIVE_TAB = {
   }
 };
 
-function createMockEndpoint(overrides = {}) {
+function createMockConnection(overrides = {}) {
   return {
     targetType: TARGET_TYPES.CAMUNDA_CLOUD,
     id: 'foo',
@@ -775,7 +768,7 @@ function createStartInstancePluginOverlay(props = {}) {
   const {
     activeTab = DEFAULT_ACTIVE_TAB,
     anchor = new MockAnchor(),
-    connectionChecker = new MockConnectionChecker(),
+    connectionManager = new MockConnectionManager(),
     deployment = new MockDeployment(),
     DeploymentConfigForm,
     deploymentConfigValidator = MockConfigValidator,
@@ -801,7 +794,7 @@ function createStartInstancePluginOverlay(props = {}) {
     <StartInstancePluginOverlay
       activeTab={ activeTab }
       anchor={ anchor }
-      connectionChecker={ connectionChecker }
+      connectionManager={ connectionManager }
       deployment={ deployment }
       DeploymentConfigForm={ DeploymentConfigForm }
       deploymentConfigValidator={ deploymentConfigValidator }
